[Feat] 가게 상세 화면 API 연동 정리
