--- MarketPlace/ViewModel/Main/CouponCreateGetVIewModel.swift
@@ -1,36 +1,0 @@
-import SwiftUI
-import Combine
-
-class CouponCreateGetViewModel: ObservableObject {
-    @Published var isAvailable: Bool = true
-    @Published var isLoading = false
-    @Published var errorMessage: String?
-    
-    let networkService = NetworkService()
-    
-    func setInitialAvailability(_ available: Bool) {
-        self.isAvailable = available
-    }
-    
-    func downloadCoupon(couponId: Int) async -> Bool {
-        isLoading = true
-        defer { isLoading = false }
-        
-        let result: NetworkResult<CommonMsgResDTO> = await networkService.request(
-            MemberCouponEndpoint.downloadCoupon(couponId: couponId)
-        )
-        switch result {
-        case .success:
-            await MainActor.run {
-                self.isAvailable = false
-            }
-            return false
-        case .failure(_, let message):
-            await MainActor.run {
-                self.errorMessage = message ?? "알 수 없는 오류"
-            }
-            return true
-        }
-    }
-}
-
